Configured DateRangePicker in HomeScreen in SearchFilter. SearchFilter is created ok is sent successfully. Started implementing communication between backend and frontend -> SERIALIZATION Problems with LocalDAte inside DateRange in Both SearchFilter and Room objects.

+++ app/src/main/java/com/example/distributed_hotel_booking/data/SearchFilter.kt
@@ -6,5 +6,5 @@ data class SearchFilter(
     val area: String,
     val numberOfGuests: Int,
     val rating : Float, //Could be Int or IntRange
-    val priceRange: Float // Could be FloatRange
+    val priceRange: Int // Could be FloatRange
 )
